✨ (themeSwitcher): Refactor ThemeSwitcher to use cookies for theme management and add ThemeToggle component
