+++ app/not-found.tsx
@@ -1,4 +1,3 @@
-"use client";
 import Link from "next/link";
 
 export default function NotFound() {
